# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/{publicId}/comments": {
      "get": {
        "description": "Returns root comments newest first, each with up to 3 preloaded replies and the thread reply count. Paginated with offset/limit; the default page is 10 roots. Accessible without authentication; a valid bearer token only fills viewerReaction.",
        "operationId": "CommentsController_listComments",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Quantos comentários-raiz pular.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Quantos comentários-raiz retornar.",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comments page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsPage"
                }
              }
            }
          },
          "400": {
            "description": "offset or limit outside the accepted range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found, or a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the comments of a video",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "description": "Creates a root comment, or a reply when parentId is given. Replies are one level deep: replying to a reply attaches the new comment to the same root. Rate limited to 5 requests per 60 s per IP.",
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty or too long body, or parentId that is not a uuid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or a draft of another channel, or parent comment not found in this video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 5 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Post a comment or a reply",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{commentId}/replies": {
      "get": {
        "description": "Returns the replies of a root comment beyond the preloaded ones, newest first, with offset/limit pagination. A reply has no replies (one level deep), so asking for them returns an empty page. Accessible without authentication; a valid bearer token only fills viewerReaction.",
        "operationId": "CommentsController_listReplies",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Quantas respostas pular — o cliente começa em 3, depois das pré-carregadas.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Quantas respostas retornar.",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replies page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepliesPage"
                }
              }
            }
          },
          "400": {
            "description": "commentId is not a uuid, or offset/limit out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found, or its video is a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the replies of a comment",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{commentId}/reaction": {
      "put": {
        "description": "Records or switches the reaction of the authenticated user on a comment or reply and returns its like count after the operation. One reaction per user per comment. There is no public dislike count. Rate limited to 60 requests per 60 s per IP.",
        "operationId": "CommentsController_setReaction",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reaction recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "commentId is not a uuid, or type is not like/dislike",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found, or its video is a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 60 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Like or dislike a comment",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "description": "Removes the reaction of the authenticated user from a comment or reply and returns its like count. Idempotent. Rate limited to 60 requests per 60 s per IP.",
        "operationId": "CommentsController_removeReaction",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reaction removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "commentId is not a uuid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found, or its video is a draft of another channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "More than 60 requests per 60 s from the same IP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove the reaction from a comment",
        "tags": [
          "comments"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SetReactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "description": "Reação do usuário: like ou dislike.",
            "enum": [
              "like",
              "dislike"
            ],
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Texto do comentário; aparado nas pontas antes de validar.",
            "minLength": 1,
            "maxLength": 2000
          },
          "parentId": {
            "type": "string",
            "description": "Comentário respondido; ausente cria um comentário-raiz.",
            "format": "uuid"
          }
        },
        "required": [
          "body"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "ReactionType": {
        "type": "string",
        "enum": [
          "like",
          "dislike"
        ],
        "description": "Reação de quem pede; null para o visitante anônimo (social-interactions-anonymous-gate/TD-02)."
      },
...
      "ReactionStateResponse": {
        "type": "object",
        "properties": {
          "viewerReaction": {
            "nullable": true,
            "description": "Reação do usuário após a operação; null quando não há.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReactionType"
              }
            ]
          },
          "likesCount": {
            "type": "number",
            "description": "Contagem de likes do alvo após a operação, lida na mesma transação."
          }
        },
        "required": [
          "viewerReaction",
          "likesCount"
        ]
      },
...
      "CommentAuthor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome de exibição do canal de quem comentou."
          },
          "nickname": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "nickname"
        ]
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Nulo para comentário-raiz; o id da raiz para resposta (profundidade 1)."
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "likesCount": {
            "type": "number"
          },
          "viewerReaction": {
            "nullable": true,
            "description": "Reação de quem pede; null para o visitante anônimo.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReactionType"
              }
            ]
          },
          "author": {
            "$ref": "#/components/schemas/CommentAuthor"
          }
        },
        "required": [
          "id",
          "parentId",
          "body",
          "createdAt",
          "likesCount",
          "viewerReaction",
          "author"
        ]
      },
      "CommentThreadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Nulo para comentário-raiz; o id da raiz para resposta (profundidade 1)."
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "likesCount": {
            "type": "number"
          },
          "viewerReaction": {
            "nullable": true,
            "description": "Reação de quem pede; null para o visitante anônimo.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReactionType"
              }
            ]
          },
          "author": {
            "$ref": "#/components/schemas/CommentAuthor"
          },
          "replies": {
            "description": "Até 3 respostas, das mais recentes para as mais antigas.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "repliesCount": {
            "type": "number",
            "description": "Total de respostas da thread."
          }
        },
        "required": [
          "id",
          "parentId",
          "body",
          "createdAt",
          "likesCount",
          "viewerReaction",
          "author",
          "replies",
          "repliesCount"
        ]
      },
      "CommentsPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentThreadResponse"
            }
          },
          "total": {
            "type": "number",
            "description": "Total de comentários-raiz do vídeo."
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ]
      },
      "RepliesPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "total": {
            "type": "number",
            "description": "Total de respostas da raiz."
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ]
      },
...
    }
  }
}